Combine some files to avoid circular dependencies
