+++ packages/apollo-shared/src/Changes/AddAndDeleteFeatureChanges.ts
@@ -8,7 +8,29 @@ import {
 import type { AnnotationFeatureSnapshot } from '@apollo-annotation/mst'
 import type { Feature } from '@apollo-annotation/schemas'
 
-import { AddFeatureChange } from './AddFeatureChange.js'
+interface SerializedAddFeatureChangeBase extends SerializedFeatureChange {
+  typeName: 'AddFeatureChange'
+}
+
+export interface AddFeatureChangeDetails {
+  addedFeature: AnnotationFeatureSnapshot
+  parentFeatureId?: string // Parent feature to where feature will be added
+  copyFeature?: boolean // Are we copying or adding a new child feature
+  allIds?: string[]
+}
+
+interface SerializedAddFeatureChangeSingle
+  extends SerializedAddFeatureChangeBase,
+    AddFeatureChangeDetails {}
+
+interface SerializedAddFeatureChangeMultiple
+  extends SerializedAddFeatureChangeBase {
+  changes: AddFeatureChangeDetails[]
+}
+
+export type SerializedAddFeatureChange =
+  | SerializedAddFeatureChangeSingle
+  | SerializedAddFeatureChangeMultiple
 
 interface SerializedDeleteFeatureChangeBase extends SerializedFeatureChange {
   typeName: 'DeleteFeatureChange'
@@ -31,6 +53,62 @@ interface SerializedDeleteFeatureChangeMultiple
 export type SerializedDeleteFeatureChange =
   | SerializedDeleteFeatureChangeSingle
   | SerializedDeleteFeatureChangeMultiple
+
+export class AddFeatureChange extends FeatureChange {
+  typeName = 'AddFeatureChange' as const
+  changes: AddFeatureChangeDetails[]
+
+  constructor(json: SerializedAddFeatureChange, options?: ChangeOptions) {
+    super(json, options)
+    this.changes = 'changes' in json ? json.changes : [json]
+  }
+  // eslint-disable-next-line @typescript-eslint/class-literal-property-style
+  get notification() {
+    return 'Feature added successfully'
+  }
+
+  toJSON(): SerializedAddFeatureChange {
+    const { assembly, changedIds, changes, typeName } = this
+    if (changes.length === 1) {
+      const [{ addedFeature, allIds, copyFeature, parentFeatureId }] = changes
+      return {
+        typeName,
+        changedIds,
+        assembly,
+        addedFeature,
+        parentFeatureId,
+        copyFeature,
+        allIds,
+      }
+    }
+    return { typeName, changedIds, assembly, changes }
+  }
+
+  getInverse() {
+    const { assembly, changedIds, changes, logger } = this
+    const inverseChangedIds = [...changedIds].reverse()
+    const inverseChanges = [...changes].reverse().map((addFeatureChange) => ({
+      deletedFeature: addFeatureChange.addedFeature,
+      parentFeatureId: addFeatureChange.parentFeatureId,
+    }))
+
+    return new DeleteFeatureChange(
+      {
+        changedIds: inverseChangedIds,
+        typeName: 'DeleteFeatureChange',
+        changes: inverseChanges,
+        assembly,
+      },
+      { logger },
+    )
+  }
+}
+
+export function isAddFeatureChange(
+  change: unknown,
+): change is AddFeatureChange {
+  return (change as AddFeatureChange).typeName === 'AddFeatureChange'
+}
 
 export class DeleteFeatureChange extends FeatureChange {
   typeName = 'DeleteFeatureChange' as const
