--- packages/apollo-shared/src/Changes/AddFeatureChange.ts
@@ -1,89 +1,0 @@
-/* eslint-disable @typescript-eslint/no-unnecessary-condition */
-import {
-  type ChangeOptions,
-  FeatureChange,
-  type SerializedFeatureChange,
-} from '@apollo-annotation/common'
-import type { AnnotationFeatureSnapshot } from '@apollo-annotation/mst'
-
-import { DeleteFeatureChange } from './DeleteFeatureChange.js'
-
-interface SerializedAddFeatureChangeBase extends SerializedFeatureChange {
-  typeName: 'AddFeatureChange'
-}
-
-export interface AddFeatureChangeDetails {
-  addedFeature: AnnotationFeatureSnapshot
-  parentFeatureId?: string // Parent feature to where feature will be added
-  copyFeature?: boolean // Are we copying or adding a new child feature
-  allIds?: string[]
-}
-
-interface SerializedAddFeatureChangeSingle
-  extends SerializedAddFeatureChangeBase,
-    AddFeatureChangeDetails {}
-
-interface SerializedAddFeatureChangeMultiple
-  extends SerializedAddFeatureChangeBase {
-  changes: AddFeatureChangeDetails[]
-}
-
-export type SerializedAddFeatureChange =
-  | SerializedAddFeatureChangeSingle
-  | SerializedAddFeatureChangeMultiple
-
-export class AddFeatureChange extends FeatureChange {
-  typeName = 'AddFeatureChange' as const
-  changes: AddFeatureChangeDetails[]
-
-  constructor(json: SerializedAddFeatureChange, options?: ChangeOptions) {
-    super(json, options)
-    this.changes = 'changes' in json ? json.changes : [json]
-  }
-  // eslint-disable-next-line @typescript-eslint/class-literal-property-style
-  get notification() {
-    return 'Feature added successfully'
-  }
-
-  toJSON(): SerializedAddFeatureChange {
-    const { assembly, changedIds, changes, typeName } = this
-    if (changes.length === 1) {
-      const [{ addedFeature, allIds, copyFeature, parentFeatureId }] = changes
-      return {
-        typeName,
-        changedIds,
-        assembly,
-        addedFeature,
-        parentFeatureId,
-        copyFeature,
-        allIds,
-      }
-    }
-    return { typeName, changedIds, assembly, changes }
-  }
-
-  getInverse() {
-    const { assembly, changedIds, changes, logger } = this
-    const inverseChangedIds = [...changedIds].reverse()
-    const inverseChanges = [...changes].reverse().map((addFeatureChange) => ({
-      deletedFeature: addFeatureChange.addedFeature,
-      parentFeatureId: addFeatureChange.parentFeatureId,
-    }))
-
-    return new DeleteFeatureChange(
-      {
-        changedIds: inverseChangedIds,
-        typeName: 'DeleteFeatureChange',
-        changes: inverseChanges,
-        assembly,
-      },
-      { logger },
-    )
-  }
-}
-
-export function isAddFeatureChange(
-  change: unknown,
-): change is AddFeatureChange {
-  return (change as AddFeatureChange).typeName === 'AddFeatureChange'
-}
